requirements.txt
flask
gunicorn
pillow
scipy
pandas
numpy
opencage
python-dotenv